apps/worker/requirements.txt
celery
redis
sqlalchemy
sentence-transformers
spacy
faiss-cpu
python-dotenv
beautifulsoup4
requests
